requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
jinja2==3.1.5
python-multipart==0.0.18
httpx==0.28.1
slowapi==0.1.9
psycopg2-binary==2.9.10
alembic==1.14.1
pytest==8.3.4
pytest-asyncio==0.24.0
ruff==0.9.4
